Fix #78 - Align home and school icons
This commit fixes an issue where the icon for the home is set 16px south
and 16px east of the location given via street address. It also resolves
an incorrect syntax for defining iconSize in home and school icons.
Lastly, the commit sets the position of school icons such that the
speech bubble which comes up when clicking on the icon is just north of
the center of the icon.