Adds images and legend
This commit adds three icons: a calendar to represent
year-round schools, a magnet to represent magnet
schools, and a scroll to represent charter schools.
This commit also adds the HTML elements for the legend
displaying these icons and their meaning at the top of
the school list.
Lastly, the symbols next to specific schools have been
changed from using letters to using the icons.