Querystring keyword to switch over to local django endpoint.
Adding the `env` querystring parameter to the url, the url being
the actual URL and NOT the fragment (everything after the `#`),
makes the endpoints to point to a local django instance running
on port 8001. The value for this querystring is non-important as
I am only checking on the existance of it.