[deploy] Fixes some salt stalk deployment issues with pip
The new pip version broke something in salt so margarita pinned
an old version. We are now pinning our project to the margarita
version that pins pip so that it works with salt. Salt was also
upgraded to 2015.5.8. The new margarita version expects our frontend
to go through a npm build using gulp and a bunch of new fancy tools
that we are not currently using. To avoid having to use all of these
tools, salt/project/web/app.sls was created to remove any mentions
of the npm bulid.